Add initial state in index
